feat(evaluator): enhance variable extraction with support for custom prefixes
